+++ src/user/user_init.c
@@ -42,6 +42,7 @@ void mjs_defaultSpec(mjSpec* spec) {
   spec->compiler.usethread = 1;
   spec->compiler.inertiafromgeom = mjINERTIAFROMGEOM_AUTO;
   spec->compiler.inertiagrouprange[1] = mjNGROUP-1;
+  spec->compiler.saveinertial = 0;
   mj_defaultLROpt(&spec->compiler.LRopt);
 
   // engine data