+++ unity/Runtime/Bindings/MjBindings.cs
@@ -5765,6 +5765,7 @@ public unsafe struct mjsCompiler_ {
   public byte fusestatic;
   public int inertiafromgeom;
   public fixed int inertiagrouprange[2];
+  public byte saveinertial;
   public int alignfree;
   public mjLROpt_ LRopt;
 }